Changes into application as per client requirements

+++ TaleOMeter/Source/APIClients/Auth/AuthClient.swift
@@ -50,16 +50,17 @@ class AuthClient {
         }
     }
     
-    static func logout(_ message: String = "") {
+    static func logout(_ message: String = "", moveToLogin: Bool = true) {
         DispatchQueue.main.async {
-//            if let cont = UIApplication.shared.windows.first?.rootViewController?.sideMenuController?.rootViewController as? UINavigationController, !(cont.viewControllers.last is LoginViewController) {
+            if moveToLogin, let cont = UIApplication.shared.windows.first?.rootViewController?.sideMenuController?.rootViewController as? UINavigationController, !(cont.viewControllers.last is LoginViewController) {
+                
                 let domain = Bundle.main.bundleIdentifier!
                 UserDefaults.standard.removePersistentDomain(forName: domain)
                 UserDefaults.standard.synchronize()
                 AudioPlayManager.shared.isMiniPlayerActive = false
                 AudioPlayManager.shared.isNonStop = false
-                
                 Login.setGusetData()
+                
 //                var contStacks = [UIViewController]()
 //                if let myobject = UIStoryboard(name: Constants.Storyboard.launch, bundle: nil).instantiateViewController(withIdentifier: "LaunchViewController") as? LaunchViewController {
 //                    contStacks.append(myobject)
@@ -68,8 +69,8 @@ class AuthClient {
 //                    contStacks.append(myobject)
 //                }
 //                cont.viewControllers = contStacks
-//                let myobject = UIStoryboard(name: Constants.Storyboard.auth, bundle: nil).instantiateViewController(withIdentifier: "LoginViewController")
-//                cont.pushViewController(myobject, animated: true)
+                let myobject = UIStoryboard(name: Constants.Storyboard.auth, bundle: nil).instantiateViewController(withIdentifier: "LoginViewController")
+                cont.pushViewController(myobject, animated: true)
                 if !message.isBlank {
                     Toast.show(message)
                 }
@@ -77,7 +78,23 @@ class AuthClient {
                     APIClient.shared.getJson("", feed: .Logout) { result in
                     }
                 }
-//            }
+            } else {
+                let domain = Bundle.main.bundleIdentifier!
+                UserDefaults.standard.removePersistentDomain(forName: domain)
+                UserDefaults.standard.synchronize()
+                AudioPlayManager.shared.isMiniPlayerActive = false
+                AudioPlayManager.shared.isNonStop = false
+                
+                Login.setGusetData()
+                
+                if !message.isBlank {
+                    Toast.show(message)
+                }
+                DispatchQueue.global(qos: .background).async {
+                    APIClient.shared.getJson("", feed: .Logout) { result in
+                    }
+                }
+            }
         }
     }
